+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python-contrib/compare/v0.200...HEAD)
 
+### Changed
+- `opentelemetry-propagator-ot-trace` Use `TraceFlags` object in `extract`
+  ([#472](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/472))
+
 ### Added
 - Move `opentelemetry-instrumentation` from core repository
   ([#465](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/465))